tests de RoomTypeService

+++ hr/hr-model/src/main/java/com/ontimize/hr/model/core/service/RoomTypeService.java
@@ -1,12 +1,15 @@
 package com.ontimize.hr.model.core.service;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Pattern;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.DuplicateKeyException;
+import org.springframework.jdbc.BadSqlGrammarException;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.stereotype.Service;
 
@@ -24,6 +27,16 @@ import com.ontimize.jee.server.dao.DefaultOntimizeDaoHelper;
 @Lazy
 public class RoomTypeService implements IRoomTypeService{
 	
+	public static final String GENERIC_ERROR = "ERROR";
+
+	public static final String DATA_ERROR = "DATA ERROR CHECK PROVIDED VALUES AND COLUMN NAMES";
+
+	public static final String GENERIC_DATA_VIOLATION_EXCEPTION = "ERROR DATA VIOLATION EXCEPTION";
+
+	public static final String TYPE_IN_USE = "CAN NOT DELETE TYPE IN USE";
+
+	public static final String NO_RECORDS_TO_DELETE = "NO RECORDS TO DELETE";
+
 	public static final String ROOM_TYPE_CAN_NOT_BE_BLANK = "ROOM TYPE CAN NOT BE BLANK";
 
 	public static final String DUPLICATED_ROOM_TYPE_NAME = "DUPLICATED ROOM TYPE NAME";
@@ -46,8 +59,12 @@ public class RoomTypeService implements IRoomTypeService{
 	@Secured({ PermissionsProviderSecured.SECURED })
 	public EntityResult roomTypeQuery(Map<String, Object> keyMap, List<String> attrList)
 			throws OntimizeJEERuntimeException {
-		
-		return this.daoHelper.query(this.roomTypeDao, keyMap, attrList);
+		try {
+			return this.daoHelper.query(this.roomTypeDao, keyMap, attrList);			
+		}
+		catch(BadSqlGrammarException e) {
+			return new EntityResultMapImpl(EntityResult.OPERATION_WRONG,12,DATA_ERROR);
+		}
 	}
 
 	/**
@@ -73,6 +90,9 @@ public class RoomTypeService implements IRoomTypeService{
 			res.setCode(EntityResult.OPERATION_WRONG);
 			res.setMessage(ROOM_TYPE_CAN_NOT_BE_BLANK);
 			return res;
+		}
+		catch(BadSqlGrammarException e) {
+			return new EntityResultMapImpl(EntityResult.OPERATION_WRONG,12,DATA_ERROR);
 		}
 	}
 
@@ -102,6 +122,9 @@ public class RoomTypeService implements IRoomTypeService{
 			res.setMessage(ROOM_TYPE_CAN_NOT_BE_BLANK);
 			return res;
 		}
+		catch (BadSqlGrammarException e) {
+			return new EntityResultMapImpl(EntityResult.OPERATION_WRONG,12,DATA_ERROR);
+		}
 	}
 
 	/**
@@ -114,7 +137,21 @@ public class RoomTypeService implements IRoomTypeService{
 	@Override
 	@Secured({ PermissionsProviderSecured.SECURED })
 	public EntityResult roomTypeDelete(Map<String, Object> keyMap) throws OntimizeJEERuntimeException {
-		return this.daoHelper.delete(this.roomTypeDao, keyMap);
+		try
+		{
+			if (this.daoHelper.query(roomTypeDao, keyMap, Arrays.asList(RoomTypeDao.ATTR_ID)).calculateRecordNumber()==0) return new EntityResultMapImpl(EntityResult.OPERATION_WRONG,12,NO_RECORDS_TO_DELETE);
+			return this.daoHelper.delete(this.roomTypeDao, keyMap);
+		}
+		catch (DataIntegrityViolationException e) {
+			if (e.getMessage()!=null && e.getMessage().contains("fk_type_room")) return new EntityResultMapImpl(EntityResult.OPERATION_WRONG,12,TYPE_IN_USE);
+			return new EntityResultMapImpl(EntityResult.OPERATION_WRONG ,12,GENERIC_DATA_VIOLATION_EXCEPTION);
+		}
+		catch (BadSqlGrammarException e) {
+//			if(e.getCause()!=null && e.getCause().getMessage()!=null && e.getCause().getMessage().contains("column"))
+//				return new EntityResultMapImpl(EntityResult.OPERATION_WRONG,12, String.format("COLUMN %s DOES NOT EXIST", Pattern.compile("«\\w+»").matcher(e.getCause().).group()));
+			return new EntityResultMapImpl(EntityResult.OPERATION_WRONG,12,DATA_ERROR);
+		}
+		
 	}
 
 }
